+++ hail/hail/test/src/is/hail/expr/ir/Aggregators2Suite.scala
@@ -645,7 +645,7 @@ class Aggregators2Suite extends HailSuite {
   }
 
   @Test def testCollectBig(): scalatest.Assertion = {
-    val seqOpArgs = Array.tabulate(100)(i => FastSeq(I64(i)))
+    val seqOpArgs = Array.tabulate(100)(i => FastSeq(I64(i.toLong)))
     assertAggEquals(
       collectAggSig(TInt64),
       FastSeq(),
